requirements.txt
Django==1.10.2
argparse==1.2.1
gunicorn==19.6.0
psycopg2==2.6.2
wsgiref==0.1.2
django-tinymce==2.4.0
